requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.7
django-allauth==0.63.6
djangorestframework==3.15.2
idna==3.7
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.1
urllib3==2.2.2
